+++ crates/core/src/commands/common.rs
@@ -14,7 +14,7 @@ use crate::{
     errors::{CommandError, HiveInitialisationError, HiveLibError},
     hive::{
         HiveLocation,
-        node::{Context, Push, SharedTarget},
+        node::{Context, Name, Push, SharedTarget},
     },
 };
 
@@ -54,8 +54,9 @@ pub async fn push(
     ]);
 
     let child = run_command_with_env(
-        &CommandArguments::new(command_string, context.modifiers)
-            .mode(crate::commands::ChildOutputMode::Nix),
+        &CommandArguments::new(command_string, context.modifiers).mode(
+            crate::commands::ChildOutputMode::Nix(Some(context.name.clone())),
+        ),
         HashMap::from([(
             "NIX_SSHOPTS".into(),
             target.create_ssh_opts(context.modifiers)?,
@@ -98,9 +99,10 @@ fn get_common_command_help(error: &CommandError) -> Option<String> {
 
 pub async fn get_hive_node_names(
     location: &HiveLocation,
+    under_name: Option<&Name>,
     modifiers: SubCommandModifiers,
 ) -> Result<Vec<String>, HiveLibError> {
-    let output = evaluate_hive_attribute(location, &EvalGoal::Names, modifiers).await?;
+    let output = evaluate_hive_attribute(location, &EvalGoal::Names, under_name, modifiers).await?;
     serde_json::from_str(&output).map_err(|err| {
         HiveLibError::HiveInitialisationError(HiveInitialisationError::ParseEvaluateError(err))
     })
@@ -112,6 +114,7 @@ pub async fn get_hive_node_names(
 pub async fn evaluate_hive_attribute(
     location: &HiveLocation,
     goal: &EvalGoal<'_>,
+    under_name: Option<&Name>,
     modifiers: SubCommandModifiers,
 ) -> Result<String, HiveLibError> {
     let attribute = match location {
@@ -152,7 +155,7 @@ pub async fn evaluate_hive_attribute(
 
     let child = run_command(
         &CommandArguments::new(command_string, modifiers)
-            .mode(crate::commands::ChildOutputMode::Nix),
+            .mode(crate::commands::ChildOutputMode::Nix(under_name.cloned())),
     )
     .await?;